chore($markdown): hoist index.js to root dir
