+++ packages/@vuepress/markdown/index.js
@@ -5,16 +5,16 @@
  */
 
 const Config = require('markdown-it-chain')
-const highlight = require('./highlight')
-const highlightLinesPlugin = require('./highlightLines')
-const preWrapperPlugin = require('./preWrapper')
-const lineNumbersPlugin = require('./lineNumbers')
-const componentPlugin = require('./component')
-const hoistScriptStylePlugin = require('./hoist')
-const convertRouterLinkPlugin = require('./link')
-const containersPlugin = require('./containers')
-const markdownSlotsContainersPlugin = require('./markdownSlotsContainers')
-const snippetPlugin = require('./snippet')
+const highlight = require('./lib/highlight')
+const highlightLinesPlugin = require('./lib/highlightLines')
+const preWrapperPlugin = require('./lib/preWrapper')
+const lineNumbersPlugin = require('./lib/lineNumbers')
+const componentPlugin = require('./lib/component')
+const hoistScriptStylePlugin = require('./lib/hoist')
+const convertRouterLinkPlugin = require('./lib/link')
+const containersPlugin = require('./lib/containers')
+const markdownSlotsContainersPlugin = require('./lib/markdownSlotsContainers')
+const snippetPlugin = require('./lib/snippet')
 const emojiPlugin = require('markdown-it-emoji')
 const anchorPlugin = require('markdown-it-anchor')
 const tocPlugin = require('markdown-it-table-of-contents')
